requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2-binary==2.9.6
sqlparse==0.4.3
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Faker==18.4.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
